fix: build errors from type refactor
