+++ src/core/custom-commands/index.ts
@@ -1,4 +1,3 @@
-export * from "./types.js";
 export * from "./parser.js";
 export * from "./loader.js";
 export * from "./executor.js";
